~ move StructureType enum out of "future" package
