+++ opc-ua-stack/stack-core/src/main/java/org/eclipse/milo/opcua/stack/core/types/enumerated/StructureType.java
@@ -1,4 +1,14 @@
-package org.eclipse.milo.opcua.stack.core.types.enumerated.future;
+/*
+ * Copyright (c) 2019 the Eclipse Milo Authors
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+
+package org.eclipse.milo.opcua.stack.core.types.enumerated;
 
 import javax.annotation.Nullable;
